+++ .github/workflows/build.yml
@@ -25,10 +25,10 @@ jobs:
       - uses: actions/checkout@v2
         
       - name: NuGet Restore
-        run: nuget restore "src/App.Metrics.Extensions.Owin.sln" -NonInteractive
+        run: nuget restore "src/NewPlatform.Flexberry.AppMetrics.Owin.sln" -NonInteractive
         
       - name: Install xunit.runner.console
         run: nuget install xunit.runner.console -Version 2.4.1 -OutputDirectory testrunner
         
       - name: Compile code
-        run:  msbuild /verbosity:quiet /p:WarningLevel=0 /p:Configuration=Debug "src/App.Metrics.Extensions.Owin.sln"
+        run:  msbuild /verbosity:quiet /p:WarningLevel=0 /p:Configuration=Debug "src/NewPlatform.Flexberry.AppMetrics.Owin.sln"
