+++ src/NewPlatform.Flexberry.AppMetrics.Owin.Core/packages.config
