+++ src/NewPlatform.Flexberry.AppMetrics.Owin/packages.config
